requirements.txt
RPi.GPIO
spidev

flask
Flask-WTF
Flask-Bootstrap4

python-vlc
youtube-dl
